+++ Content/RemakeItems/Melee/RBlightedCleaver.cs
@@ -46,7 +46,7 @@ namespace CalamityOverhaul.Content.RemakeItems.Melee
                     float adjustedItemScale = player.GetAdjustedItemScale(item);
                     Projectile.NewProjectile(source, player.MountedCenter, velocity, type
                         , (int)(damage * 0.75), knockback * 0.5f, player.whoAmI
-                        , (float)player.direction * player.gravDir, 32f, adjustedItemScale);
+                        , player.direction * player.gravDir, 32f, adjustedItemScale);
                     NetMessage.SendData(MessageID.PlayerControls, -1, -1, null, player.whoAmI);
                     return false;
                 }
